feat: release new api and architecture
this unifies the old functionality with the new api and architecture. It results in a correct
integration of files and slices and allows future extension. Sadly, this requires a break in the API
and therefore the changes are not compatible with the old version. Functionality is not fully ported
from the old version, because some of the functions were not needed or will e added in the future

BREAKING CHANGE: new fluent api and architecture